AccelerometerIngestUserData Azure Function Added
After sending an "AccelerometerIngestUserData" request from the API, this request is written by the API to the relevant Azure Storage Queue (by event).

When a message reaches to a queue that queue triggers a function and that function reads a message then
process it by its event and runs its processor".

In this case that process is "AccelerometerIngestUserDataProcess" and processor is "UserDataIngestProcessor"

"UserDataIngestProcessor" implements  "IUserDataIngestProcessor" and uses its CommandHandler(s) and QueryHandler(s) with relevant Conmmand(s) and Queries(s) which comes from AccelerometerIngest" domain.

After the "UserDataIngestProcessor" finishes its job, it returns a result to the process "AccelerometerIngestUserDataProcess" and process sends that result to the response queue.

The setting file value name/desc

- AzureWebJobsStorage      : Queue Connection
- TriggerQueue                    : Which queue triggers this function
- ResponseQueue               : Where the result goes
- ConStr                              : Db Connection String

"Other Azure Functions will be use  the same pattern"